🧹 Tidy: Widgetロジックの分離とフォームスキーマの抽出
