+++ .jules/tidy.md
@@ -189,3 +189,13 @@
 2026-02-21 - [Fat Componentの分割] LandingContent.tsxのリファクタリング
 学び: 巨大なランディングページのコンポーネントは、セクションごとに分割することで可読性が劇的に向上する。静的なコンテンツが多い場合、各セクションを独立したコンポーネントとして切り出すことで、親コンポーネントが単なるレイアウト構成要素となり、保守性が高まる。
 アクション: ページのメインコンポーネントが500行を超えた場合、セクション単位での分割を積極的に検討する。
+
+2026-02-23 - [Refactor] SleepWidgetのロジック分離とForm Schemaの抽出
+
+学び:
+- `SleepWidget` 内に複雑な計算ロジックが含まれており、可読性が低かった。`frontend/lib/sleepUtils.ts` にロジックを抽出することで、コンポーネントがUIに集中できるようになった。
+- `GrowthRecordForm` や `DiaperForm` などのフォームコンポーネントがZodスキーマを内包しており、肥大化していた。`frontend/schemas/` ディレクトリを作成し、スキーマを分離することで、バリデーションロジックの再利用性とテスト容易性が向上した。
+
+アクション:
+- 他のフォーム（例えば `FeedingForm` や `NoteForm`）でも同様にスキーマを分離する。
+- ウィジェット内のロジックが複雑になった場合は、積極的に `utils` に抽出する。
